requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
boto3==1.42.51
botocore==1.42.51
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.11
django-celery-beat==2.8.1
django-storages==1.14.6
django-timezone-field==7.2.1
idna==3.11
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
